+++ src/dbnode/encoding/m3tsz/encoder.go
@@ -21,10 +21,7 @@
 package m3tsz
 
 import (
-	"bytes"
-	"encoding/binary"
 	"errors"
-	"fmt"
 	"math"
 	"time"
 
@@ -392,257 +389,6 @@ func (enc *encoder) segment(resType resultType) ts.Segment {
 	// ref we have no ref to it anymore and if by copy then the owner should
 	// be finalizing the bytes when the segment is finalized.
 	return ts.NewSegment(head, tail, ts.FinalizeHead)
-}
-
-// TimestampEncoder encapsulates the state required for a logical stream of
-// bits that represent a stream of timestamps compressed using delta-of-delta
-type TimestampEncoder struct {
-	PrevTime       time.Time
-	PrevTimeDelta  time.Duration
-	PrevAnnotation []byte
-
-	Options encoding.Options
-
-	TimeUnit xtime.Unit
-
-	hasWrittenFirst bool // Only taken into account if using the WriteTime() API.
-}
-
-// NewTimestampEncoder creates a new TimestampEncoder.
-func NewTimestampEncoder(
-	start time.Time, timeUnit xtime.Unit, opts encoding.Options) TimestampEncoder {
-	return TimestampEncoder{
-		PrevTime: start,
-		TimeUnit: initialTimeUnit(start, timeUnit),
-		Options:  opts,
-	}
-}
-
-// WriteTime encode the timestamp using delta-of-delta compression.
-func (enc *TimestampEncoder) WriteTime(
-	stream encoding.OStream, currTime time.Time, ant ts.Annotation, timeUnit xtime.Unit) error {
-	if !enc.hasWrittenFirst {
-		if err := enc.WriteFirstTime(stream, currTime, ant, timeUnit); err != nil {
-			return err
-		}
-		enc.hasWrittenFirst = true
-		return nil
-	}
-
-	return enc.WriteNextTime(stream, currTime, ant, timeUnit)
-}
-
-// WriteFirstTime encodes the first timestamp.
-func (enc *TimestampEncoder) WriteFirstTime(
-	stream encoding.OStream, currTime time.Time, ant ts.Annotation, timeUnit xtime.Unit) error {
-	// NB(xichen): Always write the first time in nanoseconds because we don't know
-	// if the start time is going to be a multiple of the time unit provided.
-	nt := xtime.ToNormalizedTime(enc.PrevTime, time.Nanosecond)
-	stream.WriteBits(uint64(nt), 64)
-	return enc.WriteNextTime(stream, currTime, ant, timeUnit)
-}
-
-// WriteNextTime encodes the next (non-first) timestamp.
-func (enc *TimestampEncoder) WriteNextTime(
-	stream encoding.OStream, currTime time.Time, ant ts.Annotation, timeUnit xtime.Unit) error {
-	enc.writeAnnotation(stream, ant)
-	tuChanged := enc.maybeWriteTimeUnitChange(stream, timeUnit)
-
-	timeDelta := currTime.Sub(enc.PrevTime)
-	enc.PrevTime = currTime
-	if tuChanged {
-		enc.writeDeltaOfDeltaTimeUnitChanged(stream, enc.PrevTimeDelta, timeDelta)
-		// NB(xichen): if the time unit has changed, we reset the time delta to zero
-		// because we can't guarantee that dt is a multiple of the new time unit, which
-		// means we can't guarantee that the delta of delta when encoding the next
-		// data point is a multiple of the new time unit.
-		enc.PrevTimeDelta = 0
-		return nil
-	}
-	err := enc.writeDeltaOfDeltaTimeUnitUnchanged(
-		stream, enc.PrevTimeDelta, timeDelta, timeUnit)
-	enc.PrevTimeDelta = timeDelta
-	return err
-}
-
-// WriteTimeUnit writes the new time unit into the stream. It exists as a standalone method
-// so that other calls can encode time unit changes without relying on the marker scheme.
-func (enc *TimestampEncoder) WriteTimeUnit(stream encoding.OStream, timeUnit xtime.Unit) {
-	stream.WriteByte(byte(timeUnit))
-	enc.TimeUnit = timeUnit
-}
-
-// maybeWriteTimeUnitChange encodes the time unit and returns true if the time unit has
-// changed, and false otherwise.
-func (enc *TimestampEncoder) maybeWriteTimeUnitChange(stream encoding.OStream, timeUnit xtime.Unit) bool {
-	if !enc.shouldWriteTimeUnit(timeUnit) {
-		return false
-	}
-
-	scheme := enc.Options.MarkerEncodingScheme()
-	encoding.WriteSpecialMarker(stream, scheme, scheme.TimeUnit())
-	enc.WriteTimeUnit(stream, timeUnit)
-	return true
-}
-
-// shouldWriteTimeUnit determines whether we should write tu as a time unit.
-// Returns true if tu is valid and differs from the existing time unit, false otherwise.
-func (enc *TimestampEncoder) shouldWriteTimeUnit(timeUnit xtime.Unit) bool {
-	if !timeUnit.IsValid() || timeUnit == enc.TimeUnit {
-		return false
-	}
-	return true
-}
-
-// shouldWriteAnnotation determines whether we should write ant as an annotation.
-// Returns true if ant is not empty and differs from the existing annotation, false otherwise.
-func (enc *TimestampEncoder) shouldWriteAnnotation(ant ts.Annotation) bool {
-	numAnnotationBytes := len(ant)
-	if numAnnotationBytes == 0 {
-		return false
-	}
-	return !bytes.Equal(enc.PrevAnnotation, ant)
-}
-
-func (enc *TimestampEncoder) writeAnnotation(stream encoding.OStream, ant ts.Annotation) {
-	if !enc.shouldWriteAnnotation(ant) {
-		return
-	}
-
-	scheme := enc.Options.MarkerEncodingScheme()
-	encoding.WriteSpecialMarker(stream, scheme, scheme.Annotation())
-
-	var buf [binary.MaxVarintLen32]byte
-	// NB: we subtract 1 for possible varint encoding savings
-	annotationLength := binary.PutVarint(buf[:], int64(len(ant)-1))
-
-	stream.WriteBytes(buf[:annotationLength])
-	stream.WriteBytes(ant)
-	enc.PrevAnnotation = ant
-}
-
-func (enc *TimestampEncoder) writeDeltaOfDeltaTimeUnitChanged(
-	stream encoding.OStream, prevDelta, curDelta time.Duration) {
-	// NB(xichen): if the time unit has changed, always normalize delta-of-delta
-	// to nanoseconds and encode it using 64 bits.
-	dodInNano := int64(curDelta - prevDelta)
-	stream.WriteBits(uint64(dodInNano), 64)
-}
-
-func (enc *TimestampEncoder) writeDeltaOfDeltaTimeUnitUnchanged(
-	stream encoding.OStream, prevDelta, curDelta time.Duration, timeUnit xtime.Unit) error {
-	u, err := timeUnit.Value()
-	if err != nil {
-		return err
-	}
-
-	deltaOfDelta := xtime.ToNormalizedDuration(curDelta-prevDelta, u)
-	tes, exists := enc.Options.TimeEncodingSchemes()[timeUnit]
-	if !exists {
-		return fmt.Errorf("time encoding scheme for time unit %v doesn't exist", timeUnit)
-	}
-
-	if deltaOfDelta == 0 {
-		zeroBucket := tes.ZeroBucket()
-		stream.WriteBits(zeroBucket.Opcode(), zeroBucket.NumOpcodeBits())
-		return nil
-	}
-
-	buckets := tes.Buckets()
-	for i := 0; i < len(buckets); i++ {
-		if deltaOfDelta >= buckets[i].Min() && deltaOfDelta <= buckets[i].Max() {
-			stream.WriteBits(buckets[i].Opcode(), buckets[i].NumOpcodeBits())
-			stream.WriteBits(uint64(deltaOfDelta), buckets[i].NumValueBits())
-			return nil
-		}
-	}
-	defaultBucket := tes.DefaultBucket()
-	stream.WriteBits(defaultBucket.Opcode(), defaultBucket.NumOpcodeBits())
-	stream.WriteBits(uint64(deltaOfDelta), defaultBucket.NumValueBits())
-	return nil
-}
-
-func initialTimeUnit(start time.Time, tu xtime.Unit) xtime.Unit {
-	tv, err := tu.Value()
-	if err != nil {
-		return xtime.None
-	}
-	// If we want to use tu as the time unit for start, start must
-	// be a multiple of tu.
-	startInNano := xtime.ToNormalizedTime(start, time.Nanosecond)
-	tvInNano := xtime.ToNormalizedDuration(tv, time.Nanosecond)
-	if startInNano%tvInNano == 0 {
-		return tu
-	}
-	return xtime.None
-}
-
-// IntSigBitsTracker is used to track the number of significant bits
-// which should be used to encode the delta between two integers.
-type IntSigBitsTracker struct {
-	NumSig             uint8 // current largest number of significant places for int diffs
-	CurHighestLowerSig uint8
-	NumLowerSig        uint8
-}
-
-// WriteIntValDiff writes the provided val diff bits along with
-// whether the bits are negative or not.
-func (t *IntSigBitsTracker) WriteIntValDiff(
-	stream encoding.OStream, valBits uint64, neg bool) {
-	if neg {
-		stream.WriteBit(opcodeNegative)
-	} else {
-		stream.WriteBit(opcodePositive)
-	}
-
-	stream.WriteBits(valBits, int(t.NumSig))
-}
-
-// WriteIntSig writes the number of significant bits of the diff if it has changed and
-// updates the IntSigBitsTracker.
-func (t *IntSigBitsTracker) WriteIntSig(stream encoding.OStream, sig uint8) {
-	if t.NumSig != sig {
-		stream.WriteBit(opcodeUpdateSig)
-		if sig == 0 {
-			stream.WriteBit(OpcodeZeroSig)
-		} else {
-			stream.WriteBit(OpcodeNonZeroSig)
-			stream.WriteBits(uint64(sig-1), NumSigBits)
-		}
-	} else {
-		stream.WriteBit(opcodeNoUpdateSig)
-	}
-
-	t.NumSig = sig
-}
-
-// TrackNewSig gets the new number of significant bits given the
-// number of significant bits of the current diff. It takes into
-// account thresholds to try and find a value that's best for the
-// current data
-func (t *IntSigBitsTracker) TrackNewSig(numSig uint8) uint8 {
-	newSig := t.NumSig
-
-	if numSig > t.NumSig {
-		newSig = numSig
-	} else if t.NumSig-numSig >= sigDiffThreshold {
-		if t.NumLowerSig == 0 {
-			t.CurHighestLowerSig = numSig
-		} else if numSig > t.CurHighestLowerSig {
-			t.CurHighestLowerSig = numSig
-		}
-
-		t.NumLowerSig++
-		if t.NumLowerSig >= sigRepeatThreshold {
-			newSig = t.CurHighestLowerSig
-			t.NumLowerSig = 0
-		}
-
-	} else {
-		t.NumLowerSig = 0
-	}
-
-	return newSig
 }
 
 type resultType int
